requirements.txt
asgiref==3.6.0
black==23.3.0
build==0.10.0
cachetools==5.3.0
certifi==2022.12.7
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.2
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-jazzmin==2.6.0
django-json-widget==1.1.1
django-model-utils==4.3.1
django-rest-extensions==0.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-extensions==0.7.1
drf-yasg==1.21.5
filelock==3.11.0
flake8==6.0.0
future==0.18.3
identify==2.5.22
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.0
pathspec==0.11.1
pip-tools==6.13.0
pipreqs==0.4.11
platformdirs==3.2.0
pluggy==1.0.0
pre-commit==3.2.2
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pyproject_api==1.5.1
pyproject_hooks==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tomli==2.0.1
tox==4.4.11
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
yarg==0.1.9
